UI development continued + success?
i had to make the tough decision of sacrificing my beloved side scrolling text. it was causing all of my other functions not to work, which would significantly slow production of Amnesia. In order to continue development with success I have to cut it for now. However, maybe I'll ask an AI to re-write everything so it can do that after I have finished the game.