success but not really
turns out im NOT that retarded and the issue I was having was that i was using names embedded in the code or whatever that cant be used as file names such as location and name. silly me. anyways, I will continue development tonight because i just wanted to fix my issue and not actually do anything. also im playing golf tonight so thatll be fun. if I win, ill work extra hard and maybe try to implemennt a golf minigame.