introCutscene is done! and also combat tutorial?
finally done with a page other than the  menu. I have got both dialogue routes down, and I think the combat tutorial is done aswell. Now, I just need to animate everything in the second part of the intro cutscene and add the movement mechanics and then it should be smooth sailing! i hope, theres only 29 days left of this project. not enough.