DONE!
well, the time has come. i finished the game. It turned out to be significantly shorter than I wanted it to be, but this update I added double the amount of endings I wanted, so if you play through the whole game, it should be pretty long. Anyways, heres the list of all the changes:

-Fixed goblins morale damage not being affected by their own morale
-Buffed "Defend" in ForestFight
-Added Coward Ending
-Added Hero Ending
-Added Villain Ending
-Added Sacrifice Ending
-Fixed sessionStorage issues regarding the ending

and thats about all of it. For the rest of the time until the project is due I will be playtesting and getting reviews on it to further improve it. thank you for reading